# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ansa API
  version: 0.1.0
  description: |
    Modular-monolith HTTP API for the ansa ecosystem (B-Koda Limited).
    Shop v1 slice: merchant catalog (create shop/products, public published product read).
    Payments, escrow, and delivery are not exposed.
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:5000
    description: Local
paths:
  /health:
    get:
      tags: [Health]
      summary: Liveness (no database)
      operationId: healthRoot
      responses:
        "200":
          description: Process is up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /v1/health:
    get:
      tags: [Health]
      summary: Liveness (no database)
      operationId: healthV1
      responses:
        "200":
          description: Process is up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /v1/ready:
    get:
      tags: [Health]
      summary: Readiness (requires PostgreSQL)
      operationId: ready
      responses:
        "200":
          description: Database reachable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
        "503":
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/auth/register:
    post:
      tags: [Auth]
      summary: Create an account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/auth/login:
    post:
      tags: [Auth]
      summary: Sign in with email or phone
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token and issue a new access token
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New tokens
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/auth/logout:
    post:
      tags: [Auth]
      summary: Revoke a refresh token
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Revoked (idempotent)
  /v1/auth/me:
    get:
      tags: [Auth]
      summary: Current user
      operationId: me
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/me/shop:
    post:
      tags: [Shop]
      summary: Create the merchant shop (one per account)
      operationId: createShop
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShopRequest"
      responses:
        "201":
          description: Shop created
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
    get:
      tags: [Shop]
      summary: Get the authenticated merchant's shop
      operationId: getMyShop
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Shop
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: No shop yet
    patch:
      tags: [Shop]
      summary: Update the merchant shop
      operationId: updateMyShop
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShopRequest"
      responses:
        "200":
          description: Updated
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/me/shop/products:
    post:
      tags: [Shop]
      summary: Create a product
      operationId: createProduct
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Merchant has no shop
    get:
      tags: [Shop]
      summary: List the merchant's products (all statuses)
      operationId: listMyProducts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Product list
  /v1/me/shop/products/{productId}:
    get:
      tags: [Shop]
      summary: Get one of the merchant's products
      operationId: getMyProduct
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/productId"
      responses:
        "200":
          description: Product
        "404":
          description: Not found
    patch:
      tags: [Shop]
      summary: Update a product
      operationId: updateMyProduct
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/productId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Updated
    delete:
      tags: [Shop]
      summary: Archive a product (soft delete)
      operationId: archiveMyProduct
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/productId"
      responses:
        "200":
          description: Archived
  /v1/shops/{shopSlug}:
    get:
      tags: [Shop]
      summary: Public shop profile
      operationId: getPublicShop
      parameters:
        - $ref: "#/components/parameters/shopSlug"
      responses:
        "200":
          description: Shop
        "404":
          description: Not found
  /v1/shops/{shopSlug}/products:
    get:
      tags: [Shop]
      summary: List published products for a shop
      operationId: listPublicProducts
      parameters:
        - $ref: "#/components/parameters/shopSlug"
      responses:
        "200":
          description: Published products only
  /v1/shops/{shopSlug}/products/{productSlug}:
    get:
      tags: [Shop]
      summary: Retrieve a published product
      operationId: getPublicProduct
      parameters:
        - $ref: "#/components/parameters/shopSlug"
        - $ref: "#/components/parameters/productSlug"
      responses:
        "200":
          description: Published product
        "404":
          description: Missing, draft, or archived
components:
  parameters:
    productId:
      name: productId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    shopSlug:
      name: shopSlug
      in: path
      required: true
      schema:
        type: string
    productSlug:
      name: productSlug
      in: path
      required: true
      schema:
        type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ValidationError:
      description: Invalid request body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Email or phone already registered
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    HealthResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            status:
              type: string
              enum: [ok]
            service:
              type: string
              example: ansa-api
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details: {}
            requestId:
              type: string
    RegisterRequest:
      type: object
      required: [password]
      properties:
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          minLength: 8
    LoginRequest:
      type: object
      required: [password]
      properties:
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            user:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                ansaId:
                  type: string
                  format: uuid
                email:
                  type: string
                  nullable: true
                phone:
                  type: string
                  nullable: true
                createdAt:
                  type: string
                  format: date-time
            tokens:
              type: object
              properties:
                accessToken:
                  type: string
                refreshToken:
                  type: string
                expiresIn:
                  type: string
                  example: 15m
    CreateShopRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
    UpdateShopRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
    CreateProductRequest:
      type: object
      required: [title, priceKobo]
      properties:
        title:
          type: string
        description:
          type: string
        priceKobo:
          type: integer
          minimum: 0
          description: Price in kobo (NGN minor units). 150000 = ₦1,500.00
        status:
          type: string
          enum: [draft, published]
          default: published
        slug:
          type: string
    UpdateProductRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        priceKobo:
          type: integer
        status:
          type: string
          enum: [draft, published, archived]
        slug:
          type: string